rename : 패키지 설정 및 설정을 위한 빈 클래스 생성
