+++ src/main/java/bridge/controller/BridgeController.java
@@ -1,0 +1,4 @@
+package bridge.controller;
+
+public class BridgeController {
+}
