+++ src/main/java/bridge/handler/check/CheckHandler.java
@@ -1,0 +1,4 @@
+package bridge.handler.check;
+
+public class CheckHandler {
+}
