+++ src/main/java/bridge/handler/exception/ExceptionHandler.java
@@ -1,0 +1,4 @@
+package bridge.handler.exception;
+
+public class ExceptionHandler {
+}
