+++ src/main/java/bridge/repository/BridgeMaker.java
@@ -1,6 +1,8 @@
-package bridge;
+package bridge.repository;
 
 import java.util.List;
+
+import bridge.BridgeNumberGenerator;
 
 /**
  * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
